FE-Afi-02: Edit Verification, Update login button
