+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.12.1",
+    "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.60",
     "@material-ui/utils": "^4.11.2",
@@ -24,11 +24,11 @@
     "react-responsive-carousel": "^3.2.19",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "react-swipeable-views": "^0.14.0",
+    "react-toastify": "^7.0.4",
     "reactstrap": "^8.9.0",
     "redux": "^4.1.0",
     "redux-thunk": "^2.3.0",
-    "react-swipeable-views": "^0.14.0",
-    "react-toastify": "^7.0.4",
     "styled-components": "^5.3.0",
     "web-vitals": "^1.1.2"
   },
