+++ src/pages/landing/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import CarouselComp from '../../components/carousel';
-import NavbarComp from '../../components/navbar';
 
 class LandingPage extends React.Component {
     constructor(props) {
