+++ src/pages/profile/index.js
@@ -3,7 +3,6 @@ import { injectStyle } from "react-toastify/dist/inject-style";
 import PropTypes from "prop-types";
 import SwipeableViews from "react-swipeable-views";
 import { useTheme } from "@material-ui/core/styles";
-import NavbarComp from "../../components/navbar";
 import { Container } from "react-bootstrap";
 import ProfileBox from "../../components/profileBox";
 import AddressBox from "../../components/addressBox";
@@ -70,7 +69,6 @@ const UserProfile = () => {
 
   return (
     <div>
-      <NavbarComp />
       <Container>
       <Box sx={{ bgcolor: 'background.paper' }}>
         <AppBar position="static">
